move web interface shutdown related logic to its own file
